Updated Game Controller
Game controller now has startNewGame which delegates to the decoding board
as well as keeps track of what current step the player is on.

Added more test around calcuating HintPegs. The logic to check as to who
won the game should be discussed further. ( keep it in the GameController
or push it up to the GameDriver)